Add files via upload
You can view the webpages in files index1.html to index9.html, and the remaining files uploaded are images and icons that I have used in the project. A mobile.html file is a mobile site You can view it on mobile. Download and extract the files in Visual Studio Code, and run the HTML file through an open live server where you can see the webpage.